Mobile Cloud Computing/Final Project/backend/requirements.txt
Flask==0.12.2
flask-restful
pyrebase
numpy
firebase-admin
pillow==4.3.0
gunicorn==19.7.1
google.cloud
